🍄 Pop up message for edit translations (#439)
* adding new page edittranslaitons

* adding private route

* fixed grids and css for edit translations page

* stash

* hopefully fixed foot and spacing

* looking nice

* fixing spelling

* adding translation form row component

* Remove space
